refactor(web): extract MetadataSettingsTab from Settings.tsx
Final PR splitting Settings.tsx. Moves Metadata tab into its own
component. Settings logic remains in Settings.tsx (passed as props).
Depends on FE-4 and FE-5.
